LIBPERF: Split libperf source into separated modules
