+++ src/tools/perf/lib/libperf_int.h
@@ -29,6 +29,12 @@ BEGIN_C_DECLS
 #define UCT_PERF_TEST_AM_ID  5
 #define ADDR_BUF_SIZE        2048
 
+#define UCX_PERF_TEST_FOREACH(perf) \
+    while (!ucx_perf_context_done(perf))
+
+#define rte_call(_perf, _func, ...) \
+    ((_perf)->params.rte->_func((_perf)->params.rte_group, ## __VA_ARGS__))
+
 
 typedef struct ucx_perf_context        ucx_perf_context_t;
 typedef struct uct_peer                uct_peer_t;
@@ -109,7 +115,6 @@ struct ucx_perf_context {
     };
 };
 
-
 struct ucx_perf_thread_context {
     pthread_t           pt;
     int                 tid;
@@ -118,56 +123,54 @@ struct ucx_perf_thread_context {
     ucx_perf_result_t   result;
 };
 
-
 struct uct_peer {
     uct_ep_h                     ep;
     unsigned long                remote_addr;
     uct_rkey_bundle_t            rkey;
 };
 
-
 struct ucp_perf_request {
     void                         *context;
 };
 
+typedef struct {
+    ucs_status_t (*setup)(ucx_perf_context_t *perf);
+    void         (*cleanup)(ucx_perf_context_t *perf);
+    ucs_status_t (*run)(ucx_perf_context_t *perf);
+    void         (*barrier)(ucx_perf_context_t *perf);
+} ucx_perf_funcs_t;
 
-#define UCX_PERF_TEST_FOREACH(perf) \
-    while (!ucx_perf_context_done(perf))
-
-#define rte_call(_perf, _func, ...) \
-    ((_perf)->params.rte->_func((_perf)->params.rte_group, ## __VA_ARGS__))
-
+extern ucx_perf_funcs_t ucx_perf_funcs[];
 
 void ucx_perf_test_start_clock(ucx_perf_context_t *perf);
-
-
 void uct_perf_ep_flush_b(ucx_perf_context_t *perf, int peer_index);
-
-
 void uct_perf_iface_flush_b(ucx_perf_context_t *perf);
-
-
 ucs_status_t uct_perf_test_dispatch(ucx_perf_context_t *perf);
-
-
 ucs_status_t ucp_perf_test_dispatch(ucx_perf_context_t *perf);
-
-
 void ucx_perf_calc_result(ucx_perf_context_t *perf, ucx_perf_result_t *result);
-
-
 void uct_perf_barrier(ucx_perf_context_t *perf);
-
-
 void ucp_perf_barrier(ucx_perf_context_t *perf);
 
+ucs_status_t ucp_perf_test_alloc_mem(ucx_perf_context_t *perf);
+void ucp_perf_test_free_mem(ucx_perf_context_t *perf);
+ucs_status_t uct_perf_test_alloc_mem(ucx_perf_context_t *perf);
+void uct_perf_test_free_mem(ucx_perf_context_t *perf);
+ucs_status_t ucx_perf_thread_spawn(ucx_perf_context_t *perf,
+                                   ucx_perf_result_t* result);
+void ucx_perf_test_prepare_new_run(ucx_perf_context_t *perf,
+                                   const ucx_perf_params_t *params);
+void ucx_perf_set_warmup(ucx_perf_context_t* perf,
+                         const ucx_perf_params_t* params);
+/**
+ * Get the total length of the message size given by parameters
+ */
+size_t ucx_perf_get_message_size(const ucx_perf_params_t *params);
 
 static UCS_F_ALWAYS_INLINE int ucx_perf_context_done(ucx_perf_context_t *perf)
 {
     return ucs_unlikely((perf->current.iters >= perf->max_iter) ||
                         (perf->current.time  > perf->end_time));
 }
-
 
 static inline void ucx_perf_get_time(ucx_perf_context_t *perf)
 {
@@ -210,25 +213,6 @@ static inline void ucx_perf_update(ucx_perf_context_t *perf,
 
         perf->prev = perf->current;
     }
-}
-
-
-/**
- * Get the total length of the message size given by parameters
- */
-static inline
-size_t ucx_perf_get_message_size(const ucx_perf_params_t *params)
-{
-    size_t length, it;
-
-    ucs_assert(params->msg_size_list != NULL);
-
-    length = 0;
-    for (it = 0; it < params->msg_size_cnt; ++it) {
-        length += params->msg_size_list[it];
-    }
-
-    return length;
 }
 
 END_C_DECLS
