+++ src/tools/perf/perftest_run.c
@@ -4,7 +4,20 @@
 * See file LICENSE for terms.
 */
 
+#ifdef HAVE_CONFIG_H
+#  include "config.h"
+#endif
+
 #include "perftest.h"
+
+#include <ucs/sys/string.h>
+#include <ucs/sys/sys.h>
+#include <ucs/sys/sock.h>
+#include <ucs/debug/log.h>
+
+#include <getopt.h>
+#include <string.h>
+#include <locale.h>
 
 
 void print_progress(char **test_names, unsigned num_names,
@@ -167,125 +180,55 @@ static void print_test_name(struct perftest_context *ctx)
     }
 }
 
-static void print_memory_type_usage(void)
+static ucs_status_t read_batch_file(FILE *batch_file, const char *file_name,
+                                    int *line_num, perftest_params_t *params,
+                                    char** test_name_p)
 {
-    ucs_memory_type_t it;
-    for (it = UCS_MEMORY_TYPE_HOST; it < UCS_MEMORY_TYPE_LAST; it++) {
-        if (ucx_perf_mem_type_allocators[it] != NULL) {
-            printf("                        %s - %s\n",
-                   ucs_memory_type_names[it],
-                   ucs_memory_type_descs[it]);
+#define MAX_SIZE 256
+#define MAX_ARG_SIZE 2048
+    ucs_status_t status;
+    char buf[MAX_ARG_SIZE];
+    char error_prefix[MAX_ARG_SIZE];
+    int argc;
+    char *argv[MAX_SIZE + 1];
+    int c;
+    char *p;
+
+    do {
+        if (fgets(buf, sizeof(buf) - 1, batch_file) == NULL) {
+            return UCS_ERR_NO_ELEM;
+        }
+        ++(*line_num);
+
+        argc = 0;
+        p = strtok(buf, " \t\n\r");
+        while (p && (argc < MAX_SIZE)) {
+            argv[argc++] = p;
+            p = strtok(NULL, " \t\n\r");
+        }
+        argv[argc] = NULL;
+    } while ((argc == 0) || (argv[0][0] == '#'));
+
+    ucs_snprintf_safe(error_prefix, sizeof(error_prefix),
+                      "in batch file '%s' line %d: ", file_name, *line_num);
+
+    optind = 1;
+    while ((c = getopt (argc, argv, TEST_PARAMS_ARGS)) != -1) {
+        status = parse_test_params(params, c, optarg);
+        if (status != UCS_OK) {
+            ucs_error("%s-%c %s: %s", error_prefix, c, optarg,
+                      ucs_status_string(status));
+            return status;
         }
     }
-}
 
-void usage(const struct perftest_context *ctx, const char *program)
-{
-    static const char* api_names[] = {
-        [UCX_PERF_API_UCT] = "UCT",
-        [UCX_PERF_API_UCP] = "UCP"
-    };
-    test_type_t *test;
-    int UCS_V_UNUSED rank;
-
-#ifdef HAVE_MPI
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    if (ctx->mpi && (rank != 0)) {
-        return;
+    status = adjust_test_params(params, error_prefix);
+    if (status != UCS_OK) {
+        return status;
     }
-#endif
 
-#if defined (HAVE_MPI)
-    printf("  Note: test can be also launched as an MPI application\n");
-    printf("\n");
-#elif defined (HAVE_RTE)
-    printf("  Note: this test can be also launched as an libRTE application\n");
-    printf("\n");
-#endif
-    printf("  Usage: %s [ server-hostname ] [ options ]\n", program);
-    printf("\n");
-    printf("  Common options:\n");
-    printf("     -t <test>      test to run:\n");
-    for (test = tests; test->name; ++test) {
-        printf("    %13s - %s %s\n", test->name,
-               api_names[test->api], test->desc);
-    }
-    printf("\n");
-    printf("     -s <size>      list of scatter-gather sizes for single message (%zu)\n",
-                                ctx->params.super.msg_size_list[0]);
-    printf("                    for example: \"-s 16,48,8192,8192,14\"\n");
-    printf("     -m <send mem type>[,<recv mem type>]\n");
-    printf("                    memory type of message for sender and receiver (host)\n");
-    print_memory_type_usage();
-    printf("     -n <iters>     number of iterations to run (%"PRIu64")\n", ctx->params.super.max_iter);
-    printf("     -w <iters>     number of warm-up iterations (%"PRIu64")\n",
-                                ctx->params.super.warmup_iter);
-    printf("     -c <cpulist>   set affinity to this CPU list (separated by comma) (off)\n");
-    printf("     -O <count>     maximal number of uncompleted outstanding sends\n");
-    printf("     -i <offset>    distance between consecutive scatter-gather entries (%zu)\n",
-                                ctx->params.super.iov_stride);
-    printf("     -T <threads>   number of threads in the test (%d)\n",
-                                ctx->params.super.thread_count);
-    printf("     -o             do not progress the responder in one-sided tests\n");
-    printf("     -B             register memory with NONBLOCK flag\n");
-    printf("     -b <file>      read and execute tests from a batch file: every line in the\n");
-    printf("                    file is a test to run, first word is test name, the rest of\n");
-    printf("                    the line is command-line arguments for the test.\n");
-    printf("     -R <rank>      percentile rank of the percentile data in latency tests (%.1f)\n",
-                                ctx->params.super.percentile_rank);
-    printf("     -p <port>      TCP port to use for data exchange (%d)\n", ctx->port);
-#ifdef HAVE_MPI
-    printf("     -P <0|1>       disable/enable MPI mode (%d)\n", ctx->mpi);
-#endif
-    printf("     -h             show this help message\n");
-    printf("\n");
-    printf("  Output format:\n");
-    printf("     -N             use numeric formatting (thousands separator)\n");
-    printf("     -f             print only final numbers\n");
-    printf("     -v             print CSV-formatted output\n");
-    printf("\n");
-    printf("  UCT only:\n");
-    printf("     -d <device>    device to use for testing\n");
-    printf("     -x <tl>        transport to use for testing\n");
-    printf("     -D <layout>    data layout for sender side:\n");
-    printf("                        short    - short messages (default, cannot be used for get)\n");
-    printf("                        shortiov - short io-vector messages (only for active messages)\n");
-    printf("                        bcopy    - copy-out (cannot be used for atomics)\n");
-    printf("                        zcopy    - zero-copy (cannot be used for atomics)\n");
-    printf("     -W <count>     flow control window size, for active messages (%u)\n",
-                                ctx->params.super.uct.fc_window);
-    printf("     -H <size>      active message header size (%zu), included in message size\n",
-                                ctx->params.super.uct.am_hdr_size);
-    printf("     -A <mode>      asynchronous progress mode (thread_spinlock)\n");
-    printf("                        thread_spinlock - separate progress thread with spin locking\n");
-    printf("                        thread_mutex - separate progress thread with mutex locking\n");
-    printf("                        signal - signal-based timer\n");
-    printf("\n");
-    printf("  UCP only:\n");
-    printf("     -M <thread>    thread support level for progress engine (single)\n");
-    printf("                        single     - only the master thread can access\n");
-    printf("                        serialized - one thread can access at a time\n");
-    printf("                        multi      - multiple threads can access\n");
-    printf("     -D <layout>[,<layout>]\n");
-    printf("                    data layout for sender and receiver side (contig)\n");
-    printf("                        contig - Continuous datatype\n");
-    printf("                        iov    - Scatter-gather list\n");
-    printf("     -C             use wild-card tag for tag tests\n");
-    printf("     -U             force unexpected flow by using tag probe\n");
-    printf("     -r <mode>      receive mode for stream tests (recv)\n");
-    printf("                        recv       : Use ucp_stream_recv_nb\n");
-    printf("                        recv_data  : Use ucp_stream_recv_data_nb\n");
-    printf("     -I             create context with wakeup feature enabled\n");
-    printf("     -e             create endpoints with error handling support\n");
-    printf("     -E <mode>      wait mode for tests\n");
-    printf("                        poll       : repeatedly call worker_progress\n");
-    printf("                        sleep      : go to sleep after posting requests\n");
-    printf("     -H <size>      active message header size (%zu), not included in message size\n",
-                                ctx->params.super.ucp.am_hdr_size);
-    printf("\n");
-    printf("   NOTE: When running UCP tests, transport and device should be specified by\n");
-    printf("         environment variables: UCX_TLS and UCX_[SELF|SHM|NET]_DEVICES.\n");
-    printf("\n");
+    *test_name_p = strdup(argv[0]);
+    return UCS_OK;
 }
 
 static ucs_status_t run_test_recurs(struct perftest_context *ctx,
